requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
CacheControl==0.12.11
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
dj-rest-auth==2.1.3
Django==3.1.7
django-allauth==0.44.0
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-oauth-toolkit==2.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
drf-firebase-auth==1.0.0
drf-social-oauth2==1.0.9
drf-writable-nested==0.6.3
firebase-admin==4.5.3
google-api-core==1.31.5
google-api-python-client==2.48.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.5.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.1
grpcio==1.46.3
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
jwcrypto==1.2
MarkupSafe==2.1.1
msgpack==1.0.3
oauthlib==3.2.0
packaging==21.3
proto-plus==1.20.4
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
wrapt==1.14.1
